Deduplicate crate types
Noted in the review of PR 68298, it turns out that this was previously allowed,
which could lead us to some odd (if not necessarily incorrect) behavior when
checking for the presence of crate types in the list. Deduplicating doesn't
entirely eliminate the problem -- a full fix is likely to only provide APIs
which check for "set intersection" loosely -- but this more minimal fix is
likely enough given current code, if not a 100% solution.