librustc_driver: Make --print file-names do a minor accounting of --emit.
Otherwise we print bogus output which prevents using `cargo check` with sccache,
see https://bugzilla.mozilla.org/show_bug.cgi?id=1612855#c2 for a reduced
test-case.

Ideally --print file-names will emit all the files emitted by the session, but
that seems a bit hard to do in a general way, so this is a best effort thing
which fixes the cargo check use-case.