perf: Only process changed obligations in ObligationForest
This rewrites most of ObligationForest forest so to avoid iterating
through the entire set of obligations on each `select` attempt. Instead
only the obligations that can actually make progress are processed. This
gives great speedups in benchmarks such as `inflate` which create a
large number of pending obligations.

To support this the unification tables where extended to to keep track
of which type inference variables that has actually changed at each
step. Which then lets `ObligationForest` get a list of only the changed
variables at each step which it can map back to its obligations.

In addition to this primary change, many of the other iterations in
`ObligationForest` were refactored to only process lists of the nodes
that they actually are interested in. The extra bookkeeping needed for
this was possible without the primary change but were a performance
regressions there as they slowed down the main loop. As the main loop is
no longer the main issue these optimizations could be re-applied.