perf: Let &mut T iterators forward try_fold and friends
Currently `impl Iterator for &mut T` does not forward `try_fold` to the
inner `T` which means that it is easy to lose optimized implementations
of this method and methods that forward to it.

By using specialization on `T: Sized` we can allow calls on non-trait
objects to still forward to the inner methods.