Optimize `sip::Hasher::short_write`.
This commit changes `sip::Hasher` to use the faster `short_write` approach
that was used for `SipHasher128` in #68914.

This has no effect because `sip::Hasher::short_write` is currently
unused. See the next commit for more details, and a fix.

(One difference with #68914 is that this commit doesn't apply the
`u8to64_le` change from that PR, because I found it is slower, because
it introduces `memcpy` calls with non-statically-known lengths.
Therefore, this commit also undoes the `u8to64_le` change in
`SipHasher128` for this reason. This doesn't affect `SipHasher128` much
because it doesn't use `u8to64_le` much, but I am making the change to
keep the two implementations consistent.)