Move the <dyn Error>::chain() method to a trait
Because `chain()` was only implemented for `dyn Error`, users had
to cast errors to a `&(dyn Error)` to use the `chain()` method.

```rust
    let mut iter = (&my_error as &(dyn Error)).chain();
    // or
    let mut iter = <dyn Error>::chain(&my_error);
    // or
    let mut iter = Error::chain(&my_error);
```

What I would liked to have is

```rust
pub trait Error: Debug + Display {
    fn source(&self) -> Option<&(dyn Error + 'static)> {
        None
    }

    fn chain(&self) -> Chain<'_> where Self: Sized + 'static {
        Chain {
            current: Some(self),
        }
    }
}

impl dyn Error + 'static {
    fn chain(&self) -> Chain<'_> {
        Chain {
            current: Some(self),
        }
    }
}
```

This doesn't work, because of:
https://github.com/rust-lang/rust/issues/69161

This patch introduces an ErrorChain trait, which accomplishes the
job and is implemented for `Error` and `dyn Error + 'static`.