prepare_relocation_copy: check for integer overflow in the multiplication calculation for allocation size
The allocation done by prepare_relocation_copy is of size relocations.len() * length. Usually, this kind of unsafe calculation is trivially memory corruption (allocate small chunk, and then in a wildcopy write a lot, as a function of one of the operands). A good example is https://github.com/rust-lang/rust/pull/54398/commits/4d1d1beea6020cb2b760169c215da693b4fa4d13. However, since this function is internal and I don't believe is accessible as str::repeat (with is easy as "AAAA".repeat(0x4000000000000001)), it's not qualifies as a security issue. Still, I think it's more correct and elegant to check for those stuff.

Please note that in copy_repeatedly(), which calls this function, there is a comment about that: Expects the caller to have checked bounds and alignment. However, in eval_rvalue_into_place, there is no such a check (only that length > 1 due to passing length-1). I'm not sure why it hasn't been like this in first place (maybe optimization?), so I figure out I just ask :)

Thanks!