Perform WF-check on `type`s with no type parameters
Start performing a WF-check on `type`s to verify that they will be able
to be constructed. Do *not* perform a `Sized` check, as
`type T = dyn Trait;` should be allowed. Do *not* WF-check `type`s with
type parameters because we currently lint *against*
`type T<X: Trait> = K<X>;` because we do not propagate `X: Trait`,
which means that we currently allow `type T<X> = <X as Trait>::Foo;`,
which would be rejected if we WF-checked it because it would need to be
written as `type T<X: Trait> = <X as Trait>::Foo;` to be correct.
Instead, we simply don't check it at definition and only do so at use
like we do currently. In the future, the alternatives would be to
either automatically backpropagate the `X: Trait` obligation when
WF-checking the `type` or (my preferred solution) properly propagate
the explicit `X: Trait` obligation to the uses, which would likely be a
breaking change.

Fixes #60980 by using a more appropriate span for the error.