Use underlying symbol whenever we string-ify `DefPathData`
Converting an `Ident` to a string will cause us to try to guess if the
identifier is 'raw', potentially inserting a 'r#' prefix. However, all
of the existing code that string-ifys `DefPathData` is expecting to be
working with a `Symbol`, which never has a `r#` suffix.

This was causing Rustdoc to generate links containing `r#` for
structs/traits with raw identifiers (e.g. `r#struct`), leading to broken
links when the actual html files were generated without the `r#` suffix.