Remove `Hash`, `PartialEq`, and `Eq` impls from `DefPathData`
These are a footgun, since they will compare/hash based on the `Span` of
any `Ident` in the `DefPathData`.

We now use a private `PlainDefPathData` enum as the key for the
`next_disambiguator` `HashMap`. This enum is the same as `DefPathData`,
but it stores `Symbol` instead of `Ident` (e.g. what `DefPathData` used
to store).

This create a small amount of boilerplate, but allows the rest of the
codebase to work exclusively with the normal `DefPathData` enum (which
uses `Ident`s).

All of the non-`HashMap` uses of `PartialEq` turned out to be
comparisons against a constant, so `matches!` can be used instead of
`==`.