Auto merge of #5314 - ehuss:remove-git2, r=flip1995
Remove git2 dependency.

This removes the `git2` dependency (used in the integration test).  Updating git2 is awkward because both cargo and clippy have to be updated in sync, so this removes that requirement. It didn't look like it was using the git2 library for any particular reason, so this just launches the `git` executable, which should be available more or less everywhere.

This unblocks updating Cargo.

changelog: none