Handle associated types in `const` context
When in `const` context (like array lengths) we erase all substs.
Because of this, we used to emit spurious suggestions to restrict type
parameters when using associated types on them, even though that would
never work. Instead, we now signal `const` contexts when evaluating
traits and emit a more accurate error.