WIP - Add LLVM coverage instrumentation
https://github.com/rust-lang/rust/issues/34701

Prototype coverage implementation with partial support for the various
branch types to be covered.

    cargo -vv rustc -- -Z instrument-coverage

With the `-Z instrument-coverage` flag, the compiler injects calls to a
new std::coverage module, adding counters at various points in the code.

At the end of the compile, a list of coverage regions (`Span`s) and
corresponding coverage retion IDs are printed to stdout.

When executing the program, the counters increment statically, and upon
normal exit, the counters are printed to stdout with the same
corresponding coverage region IDs.

The stdout results will be replaced by output files in a format
compatible with LLVM coverage reports.

Currently counts the following coverage regions:
    * All blocks ending in a semicolon or expression
    * break, return, and yield statements, with or without a value
      expression
    * continue statment

Not yet supported are:
    * match pattern arms without a block
    * closures without a block
    * expressions ending in '?' (that return on Result::Err)
    * lazy boolean right-hand-side expressions that may not be invoked