Promote `-Zembed-bitcode` to `-Cembed-bitcode` and default to true.
This means that bitcode embedding now occurs by default. Subsequent
commits in this PR will remove the production of compressed bitcode in
rlibs, and Cargo will soon start using this flag to avoid bitcode
embedding when it's not necessary, which is most of the time.

Also, produce an early error if `-Cembed-bitcode=no` is used with
`-Clto`.