Remove `BitcodeSection::Marker`.
It existed to minimize compile time of debug builds, but the subsequent
commits in #70458 and corresponding changes to Cargo mean that bitcode
will be omitted in most cases anyway.

As a result, we just embed full bitcode whenever embedded bitcode is
necessary.

This commit also removes `BitcodeSection`, using just a `bool` instead
now to indicate the presence of bitcode in an object file.