Add support for parsing with rust-analyzer instead of librustc_parse
This adds a new librustc_parse_ra library that implements the same
interface of librustc_parse, but uses rust-analyzer to parse the code,
and then converts the rust-analyzer AST to the rustc AST, and adds a
new -Z parse-with-rust-analyzer option that makes rustc use it, gated
by a new rust_analyzer feature flag and config.toml option.

The code is preliminary but should be essentially complete, and capable
of building large projects that only use the Rust 2018 edition (Rust 2015
is not currently supported by rust-analyzer).

Note that this is likely to accept invalid code, is not fully complete,
and provides worse error messages than the current parser, so it cannot
be stabilized without significant additional work.