Do both the non-niche layout and the niche layout and compare the result
Instead of using estimation euristic to find out if we should do one or
the other layout, always perform the non-niche layout and, when possible,
the niche layout, and compare the resulting layout.

In my opinion, this is less efficient, but was requested in the review
because it is hard to proof that the estimation is correct in every case.