Specialize io::copy for BufRead sources
Allocating an additional buffer and then copying through it is completely unnecessary if the source we're copying from is BufRead (i.e. already maintains a buffer internally). Using specialization, we can detect this and handle it accordingly.

Note that this also makes it possible to specify a custom buffer size for copy operations by wrapping the source in a BufReader::with_capacity.