disable LLD build for cross compiled hosts
At present, the LLD build requested by the --enable-lld configure option
appears to make direct use of the "llvm-config" binary built as part of
the LLVM build.  For cross built LLVM, the "llvm-config" binary produced
is intended to run on the target system, so it seems like this cannot
work as intended today.  Disable the LLD build for cross compiled hosts,
while continuing to allow --enable-lld to build a native LLD for use by
targets like Fuchsia that expect to use it for linking.