Parse unambiguous const expressions in param position
When encountering the start of an expression that cannot be a type
parameter, parse it as a const expression. For an expression to be valid
they must not contain "greater than" or "shift right" operations. The
new behavior only parses expressions that begin with a token that can
start an expression but can't start a type. Calling functions and
methods will cause a type path parsing error.