Use `Fields` everywhere relevant
The idea is that `PatStack` is mostly used in the Matrix, and `Fields`
is used whenever we're close to a relevant constructor. That way, we
have a bunch of functions to go back and forth between a `Pat` and a
`Constructor` along a `Fields`. That's a simpler mental model and better
abstraction.