Add the Stream trait
This is a direct copy of the futures crate's version, which should
enable that crate to reexport std's version in the future while
preserving backwards compatibility. Like Future, this doesn't include
any of the various convenience methods - those can be added later and
for now pulled from crates like futures.