Use spin-lock to protect queue in std::sync::Once
The old implementation uses lock-free linked list. This new
implementation uses spin-lock so it will allow extra auxillary
information to be passed from the running thread to other waiting
threads.

Using spin-lock may sounds horrible but it totally fine in this
scenario: the critical region protected by this is very short
(just adding a node to the linked list); and also this is only
executed on the cold path (only if Once isn't already completed).