test/codegen: add UEFI cross-compilation test
Add a new codegen test which cross-compiles a UEFI program and verifies
that an entry-point is emitted.

The UEFI code-generation has broken in nightly before, see for instance:

    commit c1b2fd2121091df67d6a298f535b2db83e67dde5
    Merge: faccb0f07a5 6e77729ed5a
    Author: Dylan DPC <dylan.dpc@gmail.com>
    Date:   Tue May 5 01:49:44 2020 +0200

        Rollup merge of #71881 - IsaacWoods:master, r=petrochenkov

        Correctly handle UEFI targets as Windows-like when emitting sections for LLVM bitcode

This commit adds a very simple UEFI program and verifies that
cross-compilation with the built-in `x86_64-unknown-uefi` target works.
No runtime testing, nor any elaborate code-validation is done. This
simply serves as base test that verifies compilation works and an entry
point is emitted.

Usually, you would use `libcore`, but in this test we do not want to
trigger re-compilation of `libcore` for `x86_64-unknown-uefi`. Hence, we
pull in just enough of `libcore` to make the test compile.

This UEFI program is fully functional and returns success (i.e., 0). You
can run it from your UEFI BootManager or from EfiShell.

The imported UEFI definitions are explicitly kept simple. For more
elaborate use-cases you would want the full UEFI specification (see the
`r_efi` crate for a standalone import of the base UEFI specification).