[WIP] Implement shim for types that transitively derive Clone
If a type is `#[derive(Clone)]`, and all of its transitive fields
are also `#[derive(Clone)]`, we skip generating the actual `clone`
function. Instead, we generate a 'clone shim', which just acts like the
trivial `Clone` impl for a `Copy` type. We do this even if the type
itself is not `Copy`.