Use swap for base fonts in a font family
Font-display options allow a CSS author to opt in to faster loading,
in exchange for less control over the display. Font-families will
fall back to a "nearby" (in weight/style) font. It is important
that some of the base fonts, like Source Serif Pro, load correctly.
Swap should be preferred for them.