Don't use `DesugaringKind` to track desugared operators
See https://github.com/rust-lang/rust/pull/73468#issuecomment-647048588

This removes the immediate need for PR #73468, since Clippy will no
longer be confused by the precense of a new `ExpnId` 'on top of' a
macro expansion `ExpnId`.

This makes some of the 'self move diagnostics' added by PR #72389
slightly worse. I've left the operator-related diagnostics code in
place, so it should be easy for a followup PR to plug in a new way of
detecting desugared operators.